refactor(proxy): split multimodal.go and proxy.go by concern
Two more files of the proxy package over the ceiling, same package,
pure moves with their doc comments:

- multimodal.go keeps the endpoints and the passthrough pipeline
  (1071 -> 621). multimodal_multipart.go takes the multipart request
  handling (multipartPart, parseMultipartParts, multipartQuoteEscaper,
  rebuildMultipartBody, newMultipartBodyBuilder, ingestMultipartRequest);
  passthrough_sse_mask.go takes the tail buffer, the SSE error-masking
  writer and the flush writer.
- proxy.go keeps the non-streaming response path, ChatCompletions, the
  failover loop and the first-token probe (976 -> 534).
  proxy_stream_response.go takes the streaming response handling
  (handleStreamingResponse, initStreamResponse, the emit* helpers,
  applyReasoningNormalize, applyEmptyContentStrip, handleDataChunk).

Both entries leave the size allowlist. Tests are unchanged.
